+++ conf/fe.conf
@@ -66,3 +66,4 @@ mysql_service_nio_enabled = true
 # max_conn_per_user = 100
 # qe_query_timeout_second = 300
 # qe_slow_log_ms = 5000
+metadata_failure_recovery=true
