+++ fe/fe-core/src/main/java/com/starrocks/sql/optimizer/rule/transformation/PushDownPredicateScanRule.java
@@ -47,7 +47,17 @@ public class PushDownPredicateScanRule extends TransformationRule {
 
         ScalarOperator predicates = Utils.compoundAnd(lfo.getPredicate(), logicalScanOperator.getPredicate());
         ScalarRangePredicateExtractor rangeExtractor = new ScalarRangePredicateExtractor();
+
+
+        List<ScalarOperator> a = Utils.extractConjuncts(predicates);
+
+        predicates=       Utils.compoundAnd(
+                        a.stream().map(rangeExtractor::rewriteOnlyColumn).collect(Collectors.toList()));
+
         predicates = rangeExtractor.rewriteOnlyColumn(predicates);
+
+
+        //predicates = rangeExtractor.rewriteOnlyColumn(predicates);
         predicates = scalarOperatorRewriter.rewrite(predicates,
                 ScalarOperatorRewriter.DEFAULT_REWRITE_SCAN_PREDICATE_RULES);
