+++ fe/fe-core/src/test/resources/sql/optimized-plan/predicate-extract.sql
@@ -78,19 +78,19 @@ INNER JOIN (join-predicate [3: v3 = 4: v4] post-join-predicate [null])
 [sql]
 select v1 from t0 inner join t1 on v3 = v4 where (v2 = 2 or v2 = 3);
 [result]
-INNER JOIN (join-predicate [4: v4 = 3: v3] post-join-predicate [null])
-    SCAN (columns[4: v4] predicate[null])
-    EXCHANGE SHUFFLE[3]
-        SCAN (columns[1: v1, 2: v2, 3: v3] predicate[2: v2 IN (2, 3)])
+INNER JOIN (join-predicate [3: v3 = 4: v4] post-join-predicate [null])
+    SCAN (columns[1: v1, 2: v2, 3: v3] predicate[2: v2 IN (2, 3)])
+    EXCHANGE BROADCAST
+        SCAN (columns[4: v4] predicate[null])
 [end]
 
 [sql]
 select v1 from t0 inner join t1 on v3 = v4 where (v2 = 2 or v2 = 3) and (v3 = 3 or v4 = 4);
 [result]
-INNER JOIN (join-predicate [4: v4 = 3: v3 AND 3: v3 = 3 OR 4: v4 = 4] post-join-predicate [null])
-    SCAN (columns[4: v4] predicate[null])
-    EXCHANGE SHUFFLE[3]
-        SCAN (columns[1: v1, 2: v2, 3: v3] predicate[2: v2 IN (2, 3)])
+INNER JOIN (join-predicate [3: v3 = 4: v4 AND 3: v3 = 3 OR 4: v4 = 4] post-join-predicate [null])
+    SCAN (columns[1: v1, 2: v2, 3: v3] predicate[2: v2 IN (2, 3)])
+    EXCHANGE BROADCAST
+        SCAN (columns[4: v4] predicate[null])
 [end]
 
 [sql]
@@ -166,19 +166,25 @@ INNER JOIN (join-predicate [4: v4 = 3: v3] post-join-predicate [null])
 [end]
 
 [sql]
+select v1 from t0 where (v1 = 1 or v1 = 2) and v2 = v3
+[result]
+SCAN (columns[1: v1, 2: v2, 3: v3] predicate[1: v1 IN (1, 2) AND 2: v2 = 3: v3])
+[end]
+
+[sql]
+select v1 from t0 where (v1 = 1 or v1 = 2) and v2 = v1
+[result]
+SCAN (columns[1: v1, 2: v2] predicate[1: v1 IN (1, 2) AND 2: v2 = 1: v1])
+[end]
+
+[sql]
 select * from test.t0 where (v1 = 1 or v1 = 2) and (v2 = 3 or v2 = 4);
 [result]
 SCAN (columns[1: v1, 2: v2, 3: v3] predicate[1: v1 IN (1, 2) AND 2: v2 IN (3, 4)])
 [end]
 
 [sql]
- select * from test.t0 where (v1 = 1 and v2 = 2) or (v1 = 3 and v2 = 4);
+select * from test.t0 where (v1 = 1 and v2 = 2) or (v1 = 3 and v2 = 4);
 [result]
 SCAN (columns[1: v1, 2: v2, 3: v3] predicate[1: v1 = 1 AND 2: v2 = 2 OR 1: v1 = 3 AND 2: v2 = 4 AND 1: v1 IN (1, 3) AND 2: v2 IN (2, 4)])
-[end]
-
-[sql]
-select * from t0 where v1 > 1 and v1 > 5 or v2 < 6 and v2 < 8
-[result]
-SCAN (columns[1: v1, 2: v2, 3: v3] predicate[1: v1 > 5 OR 2: v2 < 6])
 [end]
